Update tests to highlight cause of issue
It seems that the cause of this bug is due to a problem that NodeJS has when trying to parse specific minified source.
The tests have been updated to show this in action, three tests are run in a nyc/mocha environment.
The first against the uninstrumented code, the second against pre-instrumented code taken from nyc's cache, the third is the pre-instrumented code run through atom-beautify to give it human readable spacing.
The tests show that the first and third tests pass while the second test should fail.  It fails because the differences in whitespace cause node to emit different exception messages.